+++ vnext/CopyHeaders.inc
@@ -1050,7 +1050,10 @@ COPYFILES_PASS0 = \
 
 COPYFILES_PASS0 = \
 	$(COPYFILES_PASS0) \
+	ReactWindows\Chakra\ChakraJsiRuntimeFactory.h, Chakra; \
+  ReactWindows\Chakra\ChakraJsiRuntimeArgs.h, Chakra; \
 	ReactCommon\jsi\jsi.h, jsi; \
 	ReactCommon\jsi\jsi-inl.h, jsi; \
+	ReactCommon\jsi\RuntimeHolder.h, jsi; \
+	ReactCommon\jsi\ScriptStore.h, jsi; \
 	ReactCommon\jsi\V8Runtime.h, jsi; \
-	ReactWindows\Chakra\ChakraRuntime.h, jsi; \
