+++ vnext/ReactWindowsCore/DevSettings.h
@@ -12,6 +12,8 @@
 #define STRING_(s) #s
 #define STRING(s) STRING_(s)
 
+namespace facebook { namespace jsi { struct RuntimeHolderLazyInit; } }
+
 namespace facebook { namespace react {
 
 enum class JSExceptionType : int32_t
@@ -26,12 +28,6 @@ struct JSExceptionInfo
   std::string exceptionMessage;
   uint32_t exceptionId;
   std::vector<std::string> callstack;
-};
-
-enum class JSIMode {
-  None,
-  ChakraJSI,
-  V8JSI
 };
 
 struct DevSettings
@@ -73,7 +69,10 @@ struct DevSettings
   /// Dispatcher for notifications about JS engine memory consumption.
   std::shared_ptr<MemoryTracker> memoryTracker;
 
-  JSIMode jsiMode;
+  /// A factory and holder of jsi::Runtime instance to be used for this react instance.
+  /// This object should in general be used only from the JS engine thread, 
+  /// unless the specific runtime implementation explicitly guarantees reentrancy.
+  std::shared_ptr<jsi::RuntimeHolderLazyInit> jsiRuntimeHolder;
 };
 
 } }
